final/market/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
cassandra-driver==3.25.0
celery==5.4.0
channels==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
cron-descriptor==1.4.5
Django==4.2.13
django-cassandra-engine==1.9.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-debug-toolbar==4.4.6
django-prometheus==2.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
flower==2.0.1
geomet==0.2.1.post1
gunicorn==23.0.0
humanize==4.10.0
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.4
pytest-cov==5.0.0
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
scylla-driver==3.28.0
six==1.16.0
sqlparse==0.5.2
tomli==2.2.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
